app/requirements.txt
flask
requests
beautifulsoup4
flask-cors
flask-socketio
python-dotenv
# gunicorn # pour le passage en prod
python-socketio[client]
eventlet